+++ src/main/java/lotto/domain/primitive/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.primitive;
 
 import java.util.Objects;
