+++ src/main/java/lotto/domain/primitive/Money.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.primitive;
 
 public class Money {
