+++ src/main/java/lotto/domain/rating/Rating.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.rating;
 
 import java.util.Arrays;
 import java.util.NoSuchElementException;
