+++ src/main/java/lotto/domain/statistics/LottoStatistics.java
@@ -1,13 +1,17 @@
-package lotto.domain;
+package lotto.domain.statistics;
 
-public class LottoStats {
+import lotto.domain.primitive.Money;
+import lotto.domain.rating.Rating;
+import lotto.domain.rating.RatingInfo;
+
+public class LottoStatistics {
     private static final String ENTER = System.lineSeparator();
     private static final String PRINT_FORMAT = "%d개 일치 (%d원) - %d개" + ENTER;
     private static final String SECOND_PRINT_FORMAT = "%d개 일치, 보너스 볼 일치 (%d원) - %d개" + ENTER;
     private RatingInfo ratingInfo;
     private final StringBuilder log = new StringBuilder();
 
-    public LottoStats(final RatingInfo ratingInfo) {
+    public LottoStatistics(final RatingInfo ratingInfo) {
         this.ratingInfo = ratingInfo;
     }
