+++ src/test/java/lotto/domain/lotto/LottoRepositoryTest.java
@@ -1,7 +1,8 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import static org.assertj.core.api.Assertions.*;
 
+import lotto.domain.lotto.LottoRepository;
 import lotto.domain.lottomachine.RandomLottoMachine;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
