+++ src/test/java/lotto/domain/primitive/LottoNumberTest.java
@@ -1,8 +1,8 @@
-package lotto.domain;
+package lotto.domain.primitive;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import lotto.domain.LottoNumber;
+import lotto.domain.primitive.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
