+++ src/test/java/lotto/domain/primitive/MoneyTest.java
@@ -1,7 +1,8 @@
-package lotto.domain;
+package lotto.domain.primitive;
 
 import static org.assertj.core.api.Assertions.*;
 
+import lotto.domain.primitive.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
