+++ src/test/java/lotto/domain/primitive/TicketTest.java
@@ -1,9 +1,9 @@
-package lotto.domain;
+package lotto.domain.primitive;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import lotto.domain.Money;
-import lotto.domain.Ticket;
+import lotto.domain.primitive.Money;
+import lotto.domain.primitive.Ticket;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
