+++ src/test/java/lotto/domain/rating/RatingInfoTest.java
@@ -1,9 +1,9 @@
-package lotto.domain;
+package lotto.domain.rating;
 
 import static org.assertj.core.api.Assertions.*;
 
-import lotto.domain.Rating;
-import lotto.domain.RatingInfo;
+import lotto.domain.rating.Rating;
+import lotto.domain.rating.RatingInfo;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
