+++ src/test/java/lotto/domain/rating/RatingTest.java
@@ -1,11 +1,11 @@
-package lotto.domain;
+package lotto.domain.rating;
 
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.Arrays;
-import lotto.domain.Lotto;
-import lotto.domain.LottoNumber;
-import lotto.domain.Rating;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.primitive.LottoNumber;
+import lotto.domain.rating.Rating;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
