+++ src/test/java/lotto/domain/statistics/LottoStatisticsTest.java
@@ -1,23 +1,23 @@
-package lotto.domain;
+package lotto.domain.statistics;
 
 import static org.assertj.core.api.Assertions.*;
 
-import lotto.domain.LottoStats;
-import lotto.domain.Money;
-import lotto.domain.Rating;
-import lotto.domain.RatingInfo;
+import lotto.domain.primitive.Money;
+import lotto.domain.rating.Rating;
+import lotto.domain.rating.RatingInfo;
+import lotto.domain.statistics.LottoStatistics;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-class LottoStatsTest {
+class LottoStatisticsTest {
     private static final String ENTER = System.lineSeparator();
     RatingInfo ratingInfo;
-    LottoStats lottoStats;
+    LottoStatistics lottoStatistics;
 
     @BeforeEach
     void setup() {
         ratingInfo = new RatingInfo();
-        lottoStats = new LottoStats(ratingInfo);
+        lottoStatistics = new LottoStatistics(ratingInfo);
     }
 
     @Test
@@ -25,7 +25,7 @@ class LottoStatsTest {
         Rating rating = Rating.FIRST;
         int count = ratingInfo.get(rating);
 
-        String actual = lottoStats.ratingToString(rating, count);
+        String actual = lottoStatistics.ratingToString(rating, count);
 
         assertThat(actual).isEqualTo("6개 일치 (2000000000원) - 0개" + ENTER);
     }
@@ -35,20 +35,20 @@ class LottoStatsTest {
         Rating rating = Rating.SECOND;
         int count = ratingInfo.get(rating);
 
-        String actual = lottoStats.ratingToString(rating, count);
+        String actual = lottoStatistics.ratingToString(rating, count);
 
         assertThat(actual).isEqualTo("5개 일치, 보너스 볼 일치 (30000000원) - 0개" + ENTER);
     }
 
     @Test
     void getEarningRate() {
-        int actual = lottoStats.getEarningRate(new Money(5000));
+        int actual = lottoStatistics.getEarningRate(new Money(5000));
         assertThat(actual).isEqualTo(0);
     }
 
     @Test
     void getTotalSum() {
         ratingInfo.update(Rating.FIRST);
-        assertThat(lottoStats.totalSum()).isEqualTo(2000000000);
+        assertThat(lottoStatistics.totalSum()).isEqualTo(2000000000);
     }
 }
