+++ src/test/java/lotto/domain/statistics/WinningLottoTest.java
@@ -1,8 +1,11 @@
-package lotto.domain;
+package lotto.domain.statistics;
 
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.Arrays;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.primitive.LottoNumber;
+import lotto.domain.statistics.WinningLotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
